+++ protected/humhub/modules/post/widgets/views/form.php
@@ -12,7 +12,7 @@ use humhub\widgets\form\ActiveForm;
 /* @var Post $post */
 ?>
 
-<?= $form->field($post, 'message', ['template' => "{input}"])->widget(RichTextField::class, [
+<?= $form->field($post, 'message')->widget(RichTextField::class, [
     'id' => 'contentForm_message' . ($wallCreateContentForm->isModal ? 'Modal' : ''),
     'form' => $form,
     'layout' => $wallCreateContentForm->isModal ? RichTextField::LAYOUT_BLOCK : RichTextField::LAYOUT_INLINE,
@@ -22,7 +22,7 @@ use humhub\widgets\form\ActiveForm;
     'disabled' => (property_exists(Yii::$app->controller, 'contentContainer') && Yii::$app->controller->contentContainer->isArchived()),
     'disabledText' => Yii::t("PostModule.base", "This space is archived."),
     'mentioningUrl' => $mentioningUrl,
-]) ?>
+])->label(false) ?>
 
 <?= WallCreateContentFormFooter::widget([
     'contentContainer' => $post->content->container,