+++ protected/humhub/modules/tour/assets/TourAsset.php
@@ -8,7 +8,10 @@
 
 namespace humhub\modules\tour\assets;
 
+use humhub\assets\DriverJsAsset;
 use humhub\components\View;
+use humhub\modules\tour\Module;
+use humhub\modules\tour\TourConfig;
 use Yii;
 use yii\helpers\Url;
 use yii\web\AssetBundle;
@@ -35,12 +38,14 @@ class TourAsset extends AssetBundle
      * @inheritdoc
      */
     public $js = [
-        'js/bootstrap-tourist.min.js',
-        'js/humhub.tour.js',
+        'js/humhub.tour.min.js',
     ];
 
     public $css = [
-        'css/bootstrap-tourist.min.css',
+    ];
+
+    public $depends = [
+        DriverJsAsset::class,
     ];
 
     /**
@@ -49,10 +54,14 @@ class TourAsset extends AssetBundle
      */
     public static function register($view)
     {
+        /** @var Module $module */
+        $module = Yii::$app->getModule('tour');
+
         $view->registerJsConfig('tour', [
             'dashboardUrl' => Url::to(['/dashboard/dashboard']),
+            'dashboardTourId' => TourConfig::TOUR_ID_DASHBOARD,
             'completedUrl' => Url::to(['/tour/tour/tour-completed']),
-            'template' => '<div class="popover tour" role="tooltip"> <div class="arrow"></div> <h3 class="popover-title"></h3> <div class="popover-content"></div> <div class="popover-navigation"> <div class="btn-group"> <button class="btn btn-sm btn-light" data-role="prev">' . Yii::t('TourModule.base', '« Prev') . '</button> <button class="btn btn-sm btn-light" data-role="next">' . Yii::t('TourModule.base', 'Next »') . '</button> <button class="btn btn-sm btn-light" data-role="pause-resume" data-pause-text="Pause" data-resume-text="Resume">Pause</button> </div> <button class="btn btn-sm btn-light" data-role="end">' . Yii::t('TourModule.base', 'End guide') . '</button> </div> </div>',
+            'driverJsOptions' => $module->driverJsOptions,
         ]);
 
         return parent::register($view);